api/requirements.txt
asgiref==3.9.1
Django==5.2.4
django-cors-headers==4.7.0
django-rest==0.8.7
djangorestframework==3.16.0
six==1.17.0
sqlparse==0.5.3
